Avoid loosing event handlers and reactivity.
To achieve that, change the way the svg element is handled. Instead of
inlining the styles directly in the source svg and then restoring it as
it was before, make a copy of the svg in a detached element, and copy
the applied styles from the source to the target. After that remove
ignored elements. This way, we avoid changing the source svg and
don't need to restore it.